Solver - insertErrorVar look up on incorrect value
Ran into a type error here when porting, it is fetching from a Map of
Constraint:Set and instead was passing in an AbstractVariable. This
seems accidental - if you notice what would happen is every single time
a new set is created (because the result is never found), this also
probably makes garbage as a side effect.